+++ packages/sdk/src/wasm/index.ts
@@ -8,4 +8,5 @@ export { generateBakerKeys } from './accountHelpers.js';
 export * from './HdWallet.js';
 export * from './identity.js';
 export * from './credentialDeploymentTransactions.js';
-export * from './web3Id.js';
+export * from './VerifiablePresentation.js';
+export * from './VerifiablePresentationV1/index.js';